requirements.txt
asgiref
Django
pillow
sqlparse
django-crispy-forms
requests
django-multiupload
crispy-tailwind
django-filter
six